service that gets number of boxes by invoiceId

+++ src/main/java/rs/prod/blinker/entity/Product.java
@@ -29,6 +29,6 @@ public class Product extends Auditable {
     @Column(name = "price")
     private Double price;
     @Column(name = "stock")
-    private Double stock;
+    private Integer stock;
 
 }
